hip: require xnack for AMD_GPU_GFX_942_APU
Kokkos allows a user to wrap a `malloc`ed pointer into
an unmanaged SharedSpace view on HIP, so ZeroMemset
needs to work. Calling hipMemset on such a pointer
is a runtime error in rocm <= 6.2.1 (and possibly later),
so provide an alternate path that uses a GPU kernel
to set the memory. For that to work, xnack must be enabled.

In the Kokkos developer meeting 2024-12-11, it was decided
that it was reasonable to require xnack for the MI300A, so
error out of HIP initialize if its not enabled.

Update the ZeroMemset test-case to use SharedSpace, as it was
always intended to.

Signed-off-by: Carl Pearson <cwpears@sandia.gov>